+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 1.0.0 - TBD
 
-- Initial version: `Unit`, `Option`, `NonEmptyList`, and `NonEmptySet`.
+- Initial version: `Unit`, `Option`, `NonEmptyList`, `NonEmptySet`, and
+  `NonEmptyMap`.